==> R/01_data_import.R <==
###############################################################################
# 01_data_import.R
# 
# This script imports preprocessed datasets generated by Python code
# and prepares them for meta-analysis in R. It loads the following files:
# - first_method_data.csv (PRIMARY ANALYSIS)
# - single_method_data.csv (SECONDARY ANALYSIS)
# - overall_data.csv (OVERALL ANALYSIS)
# - sequential_data.csv (POSITIONAL EFFECTS ANALYSIS, if available)
#
# For each dataset, it performs necessary transformations for meta-analysis:
# - Ensuring proportions are properly bounded for logit transformation
# - Computing logit-transformed proportions and their variances
# - Adding metadata for journal names, method types, etc.
#
# The prepared data is saved for use by subsequent scripts.
###############################################################################

library(tidyverse)
library(metafor)
library(meta)
library(jsonlite)

load_codebook <- function(file_path) {
  if (!requireNamespace("jsonlite", quietly = TRUE)) {
    install.packages("jsonlite")
  }
  read_json(file_path)
}

prepare_meta_data <- function(data, proportion_col = "proportion") {
  data %>%
    mutate(
      prop_trimmed = pmin(pmax(.data[[proportion_col]], 0.00001), 0.99999),
      logit_prop = log(prop_trimmed / (1 - prop_trimmed)),
      var_logit = 1 / (sample_size * prop_trimmed * (1 - prop_trimmed))
    )
}

codebook <- possibly(load_codebook, otherwise = NULL)("codebook.json")

first_method_data <- read_csv("data/for_meta/first_method_data.csv", show_col_types = FALSE) %>%
  prepare_meta_data()
write_csv(first_method_data, "data/for_r_meta/first_method_data.csv")
message(glue::glue("First-Method dataset loaded: {nrow(first_method_data)} studies"))

single_method_data <- read_csv("data/for_meta/single_method_data.csv", show_col_types = FALSE) %>%
  prepare_meta_data()
write_csv(single_method_data, "data/for_r_meta/single_method_data.csv")
message(glue::glue("Single-Method dataset loaded: {nrow(single_method_data)} studies"))

overall_data <- read_csv("data/for_meta/overall_data.csv", show_col_types = FALSE) %>%
  prepare_meta_data()
write_csv(overall_data, "data/for_r_meta/overall_data.csv")
message(glue::glue("Overall dataset loaded: {nrow(overall_data)} studies"))

sequential_data <- read_csv("data/for_meta/sequential_data.csv", show_col_types = FALSE)
message(glue::glue("Sequential dataset loaded: {nrow(sequential_data)} method-position combinations"))

if ("raw_proportion" %in% colnames(sequential_data)) {
  sequential_data <- sequential_data %>%
    prepare_meta_data("raw_proportion") %>%
    rename(
      raw_logit_prop = logit_prop,
      raw_var_logit = var_logit
    )
}

if ("adjusted_proportion" %in% colnames(sequential_data)) {
  sequential_data <- sequential_data %>%
    mutate(
      temp_data = map2(adjusted_proportion, remaining_sample, ~{
        tibble(
          proportion = .x,
          sample_size = .y
        ) %>%
          prepare_meta_data()
      }),
      adj_logit_prop = map_dbl(temp_data, ~.x$logit_prop),
      adj_var_logit = map_dbl(temp_data, ~.x$var_logit)
    ) %>%
    select(-temp_data)
  
  message("Prepared both raw and adjusted proportions for Sequential dataset")
}

write_csv(sequential_data, "data/for_r_meta/sequential_data.csv")
